feat: migrate login to use auth0 mechanism
