+++ package.json
@@ -19,7 +19,6 @@
     "@density/ui": "^21.5.0",
     "@mapbox/mapbox-gl-geocoder": "^4.5.0",
     "@material-ui/core": "^4.7.1",
-    "auth0-js": "^9.13.2",
     "axios": "^0.19.2",
     "classnames": "^2.2.6",
     "comma-number": "^2.0.1",
