+++ src/components/login/forgot-password/README.md
@@ -1,0 +1,8 @@
+# Login
+
+The login page allows a user to enter their username and password, then click a button to login.
+It's a statful component that handles its own request to the accounts api and fetches a response
+with a token.
+
+## Component Props
+(none, connected)
