+++ src/components/logout/styles.module.scss
@@ -1,0 +1,36 @@
+@import '@density/ui/variables/colors.scss';
+
+.centered {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: 18px;
+  color: $gray600;
+  user-select: none;
+}
+
+.navbarShell {
+  height: 40px;
+  width: 100%;
+  border-bottom: 1px solid $gray300;
+}
+
+.centered h1 {
+  margin: 0px;
+  padding: 0px;
+  margin-bottom: 24px;
+  margin-top: 48px;
+  font-size: 24px;
+  font-weight: normal;
+}
+
+.centered p {
+  margin: 0px;
+  padding: 0px;
+  font-size: 16px;
+}
+
+.centered a {
+  color: $blue;
+}
