+++ src/index.tsx
@@ -40,6 +40,7 @@ import createRouter from './router';
 // Import all actions required to navigate from one page to another.
 import { impersonateUnset } from './rx-actions/impersonate';
 import routeTransitionLogin from './rx-actions/route-transition/login';
+import routeTransitionLoginForgotPassword from './rx-actions/route-transition/login-forgot-password';
 import routeTransitionLogout from './rx-actions/route-transition/logout';
 import routeTransitionSpaces from './rx-actions/route-transition/spaces';
 import routeTransitionSpacesSpace from './rx-actions/route-transition/spaces-space';
@@ -146,6 +147,7 @@ trackHashChange();
 // Create a router to listen to the store and dispatch actions when the hash changes.
 // Uses conduit, an open source router we made at Density: https://github.com/DensityCo/conduit
 const router = createRouter();
+router.addRoute('login/forgot-password', async () => { (rxDispatch as Any<FixInReview>)(routeTransitionLoginForgotPassword()) });
 router.addRoute('login', async () => { (rxDispatch as Any<FixInReview>)(routeTransitionLogin(null)) });
 router.addRoute('logout', async () => routeTransitionLogout(rxDispatch));
 router.addRoute('access_token=:oauth', async () => {});
@@ -246,6 +248,8 @@ async function preRouteAuthentication() {
   // If on the account registration page (the only page that doesn't require the user to be logged in)
   // then don't worry about any of this.
   } else if (
+    locationHash.startsWith("#/logout") ||
+    locationHash.startsWith("#/login/forgot-password") ||
     locationHash.startsWith("#/account/register") ||
     locationHash.startsWith("#/account/forgot-password")
   ) {
