+++ src/rx-actions/route-transition/login-forgot-password.ts
@@ -1,0 +1,5 @@
+export const ROUTE_TRANSITION_LOGIN_FORGOT_PASSWORD = 'ROUTE_TRANSITION_LOGIN_FORGOT_PASSWORD';
+
+export default function routeTransitionLoginForgotPassword() {
+  return { type: ROUTE_TRANSITION_LOGIN_FORGOT_PASSWORD };
+}
