+++ src/rx-actions/route-transition/logout.ts
@@ -1,29 +1,5 @@
-import sessionTokenUnset from '../session-token/unset';
-import collectionSpacesSet from '../collection/spaces-legacy/set';
-import collectionSpaceHierarchySet from '../collection/space-hierarchy/set';
-import collectionTokensSet from '../collection/tokens/set';
-import collectionWebhooksSet from '../collection/webhooks/set';
-import collectionServicesSet from '../collection/services/set';
-import { UserActionTypes } from '../../types/users';
-import accounts from '../../client/accounts';
-import { showToast } from '../toasts';
-
 export const ROUTE_TRANSITION_LOGOUT = 'ROUTE_TRANSITION_LOGOUT';
 
 export default async function routeTransitionLogout(dispatch) {
   dispatch({ type: ROUTE_TRANSITION_LOGOUT });
-  try {
-    await accounts().delete('/logout');
-    dispatch(sessionTokenUnset());
-    dispatch(collectionSpacesSet([]));
-    dispatch(collectionSpaceHierarchySet([]));
-    dispatch(collectionTokensSet([]));
-    dispatch(collectionWebhooksSet([]));
-    dispatch(collectionServicesSet([]));
-    dispatch({ type: UserActionTypes.USER_MANAGEMENT_USERS_SET, users: [] });
-    window.location.hash = '#/login';
-  } catch {
-    showToast(dispatch, { text: 'Error logging out, please try again.', type: 'error' });
-    window.location.hash = '#/account';
-  }
 }
