+++ src/rx-stores/active-page/index.ts
@@ -1,4 +1,5 @@
 import { ROUTE_TRANSITION_LOGIN } from '../../rx-actions/route-transition/login';
+import { ROUTE_TRANSITION_LOGIN_FORGOT_PASSWORD } from '../../rx-actions/route-transition/login-forgot-password';
 import { ROUTE_TRANSITION_ACCOUNT } from '../../rx-actions/route-transition/account';
 import { ROUTE_TRANSITION_LIVE_SPACE_LIST } from '../../rx-actions/route-transition/live-space-list';
 import { ROUTE_TRANSITION_LIVE_SPACE_DETAIL } from '../../rx-actions/route-transition/live-space-detail';
@@ -33,6 +34,7 @@ import createRxStore from '..';
 export enum ActivePage {
   BLANK = 'BLANK',
   LOGIN = 'LOGIN',
+  LOGIN_FORGOT_PASSWORD = 'LOGIN_FORGOT_PASSWORD',
   LOGOUT = 'LOGOUT',
   LIVE_SPACE_LIST = 'LIVE_SPACE_LIST',
   LIVE_SPACE_DETAIL = 'LIVE_SPACE_DETAIL',
@@ -69,6 +71,8 @@ export function activePageReducer(state: ActivePageState, action: Any<FixInRefac
   switch (action.type) {
   case ROUTE_TRANSITION_LOGIN:
     return ActivePage.LOGIN;
+  case ROUTE_TRANSITION_LOGIN_FORGOT_PASSWORD:
+    return ActivePage.LOGIN_FORGOT_PASSWORD;
 
   case ROUTE_TRANSITION_LIVE_SPACE_LIST:
     return ActivePage.LIVE_SPACE_LIST;
